+++ apps/ranger-next/pages/_document.tsx
@@ -1,8 +1,8 @@
 import type { HelmetData } from 'react-helmet'
 import { Helmet } from 'react-helmet'
+import { CDNJavascript } from '@ranger-theme/ui'
 import type { DocumentContext, DocumentInitialProps, DocumentProps } from 'next/document'
 import Document, { Head, Html, Main, NextScript } from 'next/document'
-// import Script from 'next/script'
 
 class NextDocument extends Document<DocumentProps & { deviceType: string; helmet: HelmetData }> {
   static async getInitialProps(
@@ -40,43 +40,31 @@ class NextDocument extends Document<DocumentProps & { deviceType: string; helmet
 
   render() {
     return (
-      // <Html lang="en" style={{ fontSize: this.props.deviceType === 'H5' ? '50px' : '100px' }}>
       <Html {...this.helmetHtmlAttr}>
         <Head>
           <meta name="robots" content="INDEX,FOLLOW" />
           <meta name="google" content="notranslate" />
           <link rel="preload" href="/styles/styles.css" as="style" />
           {...this.helmetHead}
+          <CDNJavascript
+            prodUrl="https://cdnjs.cloudflare.com"
+            modules={[
+              {
+                name: 'react',
+                path: '/ajax/libs/react/18.2.0/umd/react.production.min.js',
+                defer: true,
+                crossOrigin: 'anonymous'
+              },
+              {
+                name: 'react-dom',
+                path: '/ajax/libs/react-dom/18.2.0/umd/react-dom.production.min.js',
+                defer: true,
+                crossOrigin: 'anonymous'
+              }
+            ]}
+          />
         </Head>
         <body {...this.helmetBodyAttr}>
-          {/* <Script
-            id="__html_font_size__"
-            strategy="beforeInteractive"
-            dangerouslySetInnerHTML={{
-              __html: `
-              !(function(w) {
-                var t = w.document,
-                  n = t.documentElement,
-                  a = 'orientationchange' in w ? 'orientationchange' : 'resize',
-                  d = function() {
-                    var e = w.innerWidth || 1200;
-                    if (e <= 1200) {
-                      n.style.fontSize = e / 7.5 + 'px';
-                    } else {
-                      n.style.fontSize = e / 19.2 + 'px'
-                    }
-                  };
-                if (t.readyState === 'loading') d();
-                document.documentElement.addEventListener(\"touchmove\",function(event){
-                  if (event.touches.length > 1) event.preventDefault();
-                }, false);
-                t.addEventListener && (w.addEventListener(a, d, !1), 'interactive' === t.readyState || t.addEventListener('DOMContentLoaded', function() {
-                  setTimeout(d);
-                }, !1));
-              })(window);
-            `
-            }}
-          /> */}
           <Main />
           <NextScript />
         </body>