+++ packages/ranger-ui/src/CDNJavascript/index.ts
@@ -1,0 +1,2 @@
+export type { CDNJavascriptProps } from './cdnJavascript'
+export { default as CDNJavascript } from './cdnJavascript'