+++ packages/ranger-utils/src/index.ts
@@ -1,3 +1,4 @@
+export { clipboard } from './clipboard'
 export type { Options } from './cookies'
 export { destroyCookie, getItem, parseCookies, removeItem, setItem } from './cookies'
 export { events } from './events'